# Openmp for convolution, deconvolution and relu