--- a/include/caffe/layers/base_conv_layer.hpp
+++ b/include/caffe/layers/base_conv_layer.hpp
@@ -41,6 +41,9 @@ class BaseConvolutionLayer : public Layer<Dtype> {
       weights);
   void backward_cpu_bias(Dtype* bias, const Dtype* input);
 
+  void clear_weight_mt(void);
+  void sum_weight_mt(Dtype* weight_diff);
+
 #ifndef CPU_ONLY
   void forward_gpu_gemm(const Dtype* col_input, const Dtype* weights,
       Dtype* output, bool skip_im2col = false);
@@ -161,6 +164,10 @@ class BaseConvolutionLayer : public Layer<Dtype> {
 
   Blob<Dtype> col_buffer_;
   Blob<Dtype> bias_multiplier_;
+
+  int num_of_threads_;                 // openmp
+  std::vector<Dtype> col_buffer_mt_;   //  openmp
+  std::vector<Dtype> weight_diff_mt_;  // openmp
 };
 
 }  // namespace caffe