--- a/src/caffe/layers/base_conv_layer.cpp
+++ b/src/caffe/layers/base_conv_layer.cpp
@@ -6,6 +6,10 @@
 #include "caffe/util/im2col.hpp"
 #include "caffe/util/math_functions.hpp"
 
+#ifdef _OPENMP
+#include <omp.h>
+#endif
+
 namespace caffe {
 
 template <typename Dtype>
@@ -237,18 +241,47 @@ void BaseConvolutionLayer<Dtype>::Reshape(const vector<Blob<Dtype>*>& bottom,
     caffe_set(bias_multiplier_.count(), Dtype(1),
         bias_multiplier_.mutable_cpu_data());
   }
+
+  // ---- openmp ------------------------------------------
+  num_of_threads_ = 1;
+#ifdef _OPENMP  //TODO: omp_get_max_threads
+  num_of_threads_ = omp_get_max_threads();
+  if (num_of_threads_ < 1) {
+     LOG(WARNING) << "Base Conv layer: omp_get_max_threads() =" << num_of_threads_;
+     num_of_threads_ = 1;
+  }
+#endif
+
+  int col_buffer_mt_size = num_of_threads_ * col_buffer_.count();
+  int weight_diff_mt_size = num_of_threads_ * this->blobs_[0]->count();  
+ 
+  col_buffer_mt_.resize( col_buffer_mt_size );
+  weight_diff_mt_.resize( weight_diff_mt_size );
 }
 
 template <typename Dtype>
 void BaseConvolutionLayer<Dtype>::forward_cpu_gemm(const Dtype* input,
     const Dtype* weights, Dtype* output, bool skip_im2col) {
-  const Dtype* col_buff = input;
-  if (!is_1x1_) {
-    if (!skip_im2col) {
-      conv_im2col_cpu(input, col_buffer_.mutable_cpu_data());
-    }
-    col_buff = col_buffer_.cpu_data();
+
+  int tid = 0;
+#ifdef _OPENMP
+  tid = omp_get_thread_num();
+  if (tid >= num_of_threads_) {
+    LOG(FATAL) << "ConvLayer::Forward_cpu: omp_thread_num() =" << tid
+               << " > OMP_num_THREADS = " << num_of_threads_;
   }
+  tid = tid % num_of_threads_;  //  just to be sure
+#endif
+  int col_data_buffer_size = col_buffer_mt_.size()/num_of_threads_ ;
+
+  Dtype* col_buff = const_cast<Dtype*>(input);
+  if (!is_1x1_) {
+    col_buff = & col_buffer_mt_[ tid* col_data_buffer_size];
+    if (!skip_im2col) {
+      conv_im2col_cpu(input, col_buff );
+    } 
+  } 
+
   for (int g = 0; g < group_; ++g) {
     caffe_cpu_gemm<Dtype>(CblasNoTrans, CblasNoTrans, conv_out_channels_ /
         group_, conv_out_spatial_dim_, kernel_dim_,
@@ -268,7 +301,19 @@ void BaseConvolutionLayer<Dtype>::forward_cpu_bias(Dtype* output,
 template <typename Dtype>
 void BaseConvolutionLayer<Dtype>::backward_cpu_gemm(const Dtype* output,
     const Dtype* weights, Dtype* input) {
-  Dtype* col_buff = col_buffer_.mutable_cpu_data();
+
+  int tid = 0;
+#ifdef _OPENMP
+  tid = omp_get_thread_num();
+  if (tid >= num_of_threads_) {
+    LOG(FATAL) << "ConvLayer::backward_cpu_gemm: omp_thread_num() =" << tid
+               << " > OMP_num_THREADS = " << num_of_threads_;
+  }
+  tid = tid % num_of_threads_;  //  just to be sure
+#endif
+  int col_data_buffer_size = col_buffer_mt_.size()/num_of_threads_;
+  Dtype* col_buff = & col_buffer_mt_[ tid* col_data_buffer_size];
+
   if (is_1x1_) {
     col_buff = input;
   }
@@ -283,21 +328,58 @@ void BaseConvolutionLayer<Dtype>::backward_cpu_gemm(const Dtype* output,
   }
 }
 
+
+template <typename Dtype>
+void BaseConvolutionLayer<Dtype>::clear_weight_mt(void)
+{    
+  // TODO:  make it more optimal
+  unsigned int weight_diff_size = weight_diff_mt_.size() / num_of_threads_;  
+  caffe_memset( num_of_threads_ * weight_diff_size * sizeof(Dtype), 0., & weight_diff_mt_[0]);
+}
+
+
+template <typename Dtype>
+void BaseConvolutionLayer<Dtype>::sum_weight_mt(Dtype* weight_diff)
+{
+  //TODO: Optimize
+  // sum weight_diff over all threads
+  unsigned int weight_diff_size =  weight_diff_mt_.size() / num_of_threads_; 
+  for (int t = 0; t < num_of_threads_ ; ++t) {
+    for (int j = 0; j < weight_diff_size ; ++j) {
+      weight_diff[j] += weight_diff_mt_[t * weight_diff_size + j];
+    }
+  }
+}
+
 template <typename Dtype>
 void BaseConvolutionLayer<Dtype>::weight_cpu_gemm(const Dtype* input,
-    const Dtype* output, Dtype* weights) {
-  const Dtype* col_buff = input;
+    const Dtype* output) {
+  int tid = 0;
+#ifdef _OPENMP
+  tid = omp_get_thread_num();
+  if (tid >= num_of_threads_) {
+    LOG(FATAL) << "ConvLayer::weights_cpu_gemm: omp_thread_num() =" << tid
+               << " > OMP_num_THREADS = " << num_of_threads_;
+  }
+  tid = tid % num_of_threads_;  //  just to be sure
+#endif
+  Dtype* col_buff = const_cast<Dtype*>(input);
+  Dtype* weight_diff_data= & weight_diff_mt_[tid * (weight_diff_mt_.size()/num_of_threads_ )];
+
   if (!is_1x1_) {
-    conv_im2col_cpu(input, col_buffer_.mutable_cpu_data());
-    col_buff = col_buffer_.cpu_data();
+    int col_data_buffer_size = col_buffer_mt_.size()/num_of_threads_;
+    col_buff = & col_buffer_mt_[ tid* col_data_buffer_size];
+    conv_im2col_cpu(input, col_buff );
   }
   for (int g = 0; g < group_; ++g) {
     caffe_cpu_gemm<Dtype>(CblasNoTrans, CblasTrans, conv_out_channels_ / group_,
         kernel_dim_, conv_out_spatial_dim_,
         (Dtype)1., output + output_offset_ * g, col_buff + col_offset_ * g,
-        (Dtype)1., weights + weight_offset_ * g);
+        (Dtype)1., weight_diff_data + weight_offset_ * g);
   }
 }
+
+
 
 template <typename Dtype>
 void BaseConvolutionLayer<Dtype>::backward_cpu_bias(Dtype* bias,