@talda 2015-11-19
Note that mkl by itself uses multiple threads. I think in this case you will get a better result by setting the number of threads used by mkl to 1. 

You can do this by setting MKL_NUM_THREADS environment variable to 1.