@bhack 2015-11-19
/cc @borisgin