@ronghanghu 2015-11-19
IMO, if you use Intel MKL or OpenBLAS as CPU backend (not sure about Atlas), then it is already utilizing multi-core, so OpenMP doesn't seem necessary for Conv and Deconv. OpenMP can still help for ReLU, but may introduce more difficulties for debugging.