@jczaja 2015-11-20
@talda Thanks for suggesstion. I tried that but did not get better performance than when not restricting mkl threads. MKL libs are providing multithreaded version: mkl_rt.so and singlethreaded: mkl_sequential.so. Caffe is using multithreaded, but I made modification (just to observere if I can get better performance) to run against single threaded MKL. But results I got were weaker than when using multithreaded MKL.