@jczaja 2015-11-20
@ronghanghu Thanks for opinion. All three mentioned BLAS libs comes in singlethreaded as well as multithreaded editions. Here In performance result I used multithreaded MKL (mkl_rt.so), and singlethreaded OpenBLAS. As those configurations perform best with OpenMP changes being reviewed here. 

Few words about OpenBLAS tested. Result comes from single threaded, but I was able to make multithreaded build from source , for target architecture I was testing on (HASWELL). And I did not get better results that what I presented here.

Now, considering question if there is a point in optimizing convolution code while having multithreaded BLAS lib used, I checked if convolution itself gives performance boost (and it does) , and if relu gives performance boost (it does too). They both contribute to speedup of layer execution.
Why there is performance boost in convolution using openmp? Well, conv layer is iterating over batches eg. each loop iteration is processing one batch of data and then MKL/OpenBLAS/Atlas is to perform multiplication using multithreading. Perhaps for BLAS libs is not that easy to utilise full platform resources when working on single batch of data. On the other hand caffe OpenMP changes are implementing utilisation of threads so each of thread is running separate BLAS operation, also done in a way to avoid false sharing.