@jczaja 2015-12-01
@ronghanghu I realised that I did not answered fully on your questions. As for concerns about difficulty of debugging of openmp code, there are few points worth mentioning:
1) OpenMP is optional path, disabled by default . So contributors not focused on CPU performance should not be concerned with OpenMP path.
2) range of modifications is kept to minimum , so in case mentioned Relu, this is only one #pragma .. line that was added. OpenMP and sequential path share most of code path.
3) As for debugging OpenMP solution. With current approach of keeping sequential path of CPU alive and adding minimum openmp changes, I do not think there will be much problems with debugging OpenMP code, other than making sure that normal, sequential CPU path works.
**Moreover, Myself and my colleagues are more than happy to help if such a need of debugging OpenMP path , arise.**
4) Performance results (presented in this PR) and performance boost that will be a result of next OpenMP based PR (will get up to ~8 times better performance comparing to sequential path, when pooling and LRN are modified to use OpenMP) are making this PR worth consideration to be merged:)

If something in code is not clear of you have some suggestions to code , please let me know.

Looking forward to your feedback,