@jczaja 2015-12-18
I have updated PR :
- OpenMP is now used if available , if not avaialble then non-OpenMP build is made. 
- MKL usage from cmake was fixed/improved