@bhack 2015-12-18
There is a specific openblas branch at https://github.com/xianyi/OpenBLAS/tree/optimized_for_deeplearning?files=1