@jczaja 2015-12-18
@bhack Thanks for letting me know about effort of tuning openblas for deeplearning needs.